+++ lib/raxx/simple_server.ex
@@ -1,26 +1,34 @@
-defmodule Raxx.NotFound do
-  require EEx
+defmodule Raxx.SimpleServer do
+  @typedoc """
+  State of application server.
+
+  Original value is the configuration given when starting the raxx application.
+  """
+  @type state :: any()
+
+  @doc """
+  Called with a complete request once all the data parts of a body are received.
+
+  Passed a `Raxx.Request` and server configuration.
+  Note the value of the request body will be a string.
+  """
+  @callback handle_request(Raxx.Request.t(), state()) :: Raxx.Response.t()
+
   @eight_MB 8 * 1024 * 1024
 
-  template = Path.join(__DIR__, "./not_found.html.eex")
-  EEx.function_from_file(:defp, :not_found_page, template, [:example_module])
-
   defmacro __using__(options) do
-    body = not_found_page(__CALLER__.module)
     {options, []} = Module.eval_quoted(__CALLER__, options)
     maximum_body_length = Keyword.get(options, :maximum_body_length, @eight_MB)
 
     quote do
-      @impl Raxx.Server
-      def handle_request(_request, _state) do
-        Raxx.response(:not_found)
-        |> Raxx.set_header("content-type", "text/html")
-        |> Raxx.set_body(unquote(body))
-      end
+      @behaviour unquote(__MODULE__)
+      import Raxx
 
-      @impl Raxx.Server
+      @behaviour Raxx.Server
+
+      # @impl Raxx.Server
       def handle_head(request = %{body: false}, state) do
-        response = handle_request(%{request | body: ""}, state)
+        response = __MODULE__.handle_request(%{request | body: ""}, state)
 
         case response do
           %{body: true} -> raise "Incomplete response"
@@ -32,7 +40,7 @@ defmodule Raxx.NotFound do
         {[], {request, [], state}}
       end
 
-      @impl Raxx.Server
+      # @impl Raxx.Server
       def handle_data(data, {request, iodata_buffer, state}) do
         iodata_buffer = [data | iodata_buffer]
 
@@ -43,10 +51,10 @@ defmodule Raxx.NotFound do
         end
       end
 
-      @impl Raxx.Server
+      # @impl Raxx.Server
       def handle_tail([], {request, iodata_buffer, state}) do
         body = :erlang.iolist_to_binary(Enum.reverse(iodata_buffer))
-        response = handle_request(%{request | body: body}, state)
+        response = __MODULE__.handle_request(%{request | body: body}, state)
 
         case response do
           %{body: true} -> raise "Incomplete response"
@@ -54,7 +62,7 @@ defmodule Raxx.NotFound do
         end
       end
 
-      @impl Raxx.Server
+      # @impl Raxx.Server
       def handle_info(message, state) do
         require Logger
 
@@ -64,8 +72,6 @@ defmodule Raxx.NotFound do
 
         {[], state}
       end
-
-      defoverridable Raxx.Server
     end
   end
 end