+++ CHANGELOG.md
@@ -6,11 +6,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Next
 
+Added `Raxx.SimpleServer` behaviour
+Changed Raxx.Server using, fixed logger router
+
 ### Changed
 
 - `Raxx.set_header/2` raises an `ArgumentError` when setting host headers.
 - `ArgumentError` is raised instead of `RuntimeError` in cases of bad headers and body content.
 - `Raxx.set_body/2` raises an `ArgumentError` for GET and HEAD requests.
+
+### Removed
+
+- `Raxx.is_application?` use `Raxx.Server.verify_server`.
+- `Raxx.verify_application` use `Raxx.Server.verify_server`.
+- `Raxx.Server.is_implemented?` use `Raxx.Server.verify_server`.
 
 ## [0.16.1](https://github.com/CrowdHailer/raxx/tree/0.16.1) - 2018-09-19